Move dialog paths
