+++ frontend/components/dialogs/HistoryDialog.tsx
