+++ frontend/components/dialogs/TermsAndConditionsDialog.tsx
